Add domain class for sets of keys
Maybe this is not necessary, but it seems to be more
natural to have K[0], K[1], ... as domain instances
rather than carrying with (K, degree) to represent them.

Should be useful when defining simplicial complexes as well
or other topological instances: the System class should implement
only the nerve-complex specific features.